+++ tests/testthat/test-consistency-planning.R
@@ -28,8 +28,8 @@ test_that(desc = "(id: f5-v0.1.0-t4) Frequentist Poisson 5% materiality 2% error
 
 test_that(desc = "(id: f5-v0.1.0-t5) Bayesian Poisson 5% materiality 2.5% errors", {
   jfaRes <- planning(materiality = 0.05, conf.level = 0.95, expected = 0.025, likelihood = "poisson", prior = TRUE)
-  expect_equal(jfaRes[["n"]], 231)
-  expect_equal(jfaRes[["x"]], 5.78, tolerance = 0.001)
+  expect_equal(jfaRes[["n"]], 228)
+  expect_equal(jfaRes[["x"]], 5.7, tolerance = 0.001)
 })
 
 test_that(desc = "(id: f5-v0.1.0-t7) Frequentist binomial 1% materiality", {
@@ -57,7 +57,7 @@ test_that(desc = "(id: f5-v0.1.0-t10) Frequentist binomial 5% materiality 2 erro
 })
 
 test_that(desc = "(id: f5-v0.1.0-t11) Bayesian binomial 5% materiality 2.5% errors", {
-  prior <- auditPrior(method = 'uniform', likelihood = 'binomial')
+  prior <- auditPrior(method = 'none', likelihood = 'binomial')
   jfaRes <- planning(materiality = 0.05, conf.level = 0.95, expected = 0.025, likelihood = "binomial", prior = prior)
   expect_equal(jfaRes[["n"]], 220)
   expect_equal(jfaRes[["x"]], 5.5, tolerance = 0.001)
@@ -88,7 +88,7 @@ test_that(desc = "(id: f5-v0.1.0-t16) Frequentist hypergeometric 5% materiality 
 })
 
 test_that(desc = "(id: f5-v0.1.0-t17) Bayesian hypergeometric 5% materiality 2.5% errors", {
-  prior <- auditPrior(method = 'uniform', likelihood = 'hypergeometric', N.units = 1000)
+  prior <- auditPrior(method = 'none', likelihood = 'hypergeometric', N.units = 1000)
   jfaRes <- planning(materiality = 0.05, conf.level = 0.95, expected = 0.025, likelihood = "hypergeometric", N.units = 1000, prior = prior)
   expect_equal(jfaRes[["n"]], 182)
   expect_equal(jfaRes[["x"]], 5, tolerance = 0.001)
@@ -131,7 +131,7 @@ test_that(desc = "(id: f5-v0.1.0-t23) Frequentist poisson 3% materiality 1.3% er
 # jfa version 0.2.0
 
 test_that(desc = "(id: f5-v0.2.0-t1) Bayesian poisson 3% materiality 1.3% errors 5% min precision standard prior", {
-  prior <- auditPrior(method = 'uniform', likelihood = 'binomial')
+  prior <- auditPrior(method = 'none', likelihood = 'binomial')
   jfaRes <- planning(materiality = 0.03, conf.level = 0.95, expected = 0.013, min.precision = 0.05, likelihood = "binomial", N.units = 1000, prior = prior)
   expect_equal(jfaRes[["n"]], 293)
   expect_equal(jfaRes[["x"]], 3.81, tolerance = 0.001)
@@ -139,14 +139,14 @@ test_that(desc = "(id: f5-v0.2.0-t1) Bayesian poisson 3% materiality 1.3% errors
 
 test_that(desc = "(id: f5-v0.2.0-t2) Bayesian poisson 100% materiality 0% errors 2% min precision standard prior", {
   jfaRes <- planning(conf.level = 0.95, min.precision = 0.02, likelihood = "poisson", prior = TRUE)
-  expect_equal(jfaRes[["n"]], 150)
+  expect_equal(jfaRes[["n"]], 149)
   expect_equal(jfaRes[["x"]], 0)
 })
 
 test_that(desc = "(id: f5-v0.2.0-t3) Bayesian poisson 100% materiality 1% errors 2% min precision standard prior", {
   jfaRes <- planning(conf.level = 0.95, min.precision = 0.02, expected = 0.01, likelihood = "poisson", prior = TRUE)
-  expect_equal(jfaRes[["n"]], 220)
-  expect_equal(jfaRes[["x"]], 2.2, tolerance = 0.001)
+  expect_equal(jfaRes[["n"]], 219)
+  expect_equal(jfaRes[["x"]], 2.19, tolerance = 0.001)
 })
 
 # jfa version 0.3.0 - 0.3.1
@@ -155,27 +155,29 @@ test_that(desc = "(id: f5-v0.2.0-t3) Bayesian poisson 100% materiality 1% errors
 # jfa version 0.4.0
 
 test_that(desc = "(id: f5-v0.4.0-t1) Expected Bayes factors for zero expected errors", {
-  jfaRes <- planning(conf.level = 0.95, materiality = 0.02, expected = 0, likelihood = "poisson", prior = TRUE)
+  prior <- auditPrior(method = 'strict', likelihood = 'poisson')
+  jfaRes <- planning(conf.level = 0.95, materiality = 0.02, expected = 0, likelihood = "poisson", prior = prior)
   expect_equal(jfaRes[["posterior"]][["hypotheses"]]$bf.hmin, 19.08554, tolerance = 0.001)
   
-  prior <- auditPrior(method = 'uniform', likelihood = 'binomial')
+  prior <- auditPrior(method = 'none', likelihood = 'binomial')
   jfaRes <- planning(conf.level = 0.95, materiality = 0.02, expected = 0, likelihood = "binomial", prior = prior)
   expect_equal(jfaRes[["posterior"]][["hypotheses"]]$bf.hmin, 945.2848, tolerance = 0.001)
   
-  prior <- auditPrior(method = 'uniform', likelihood = 'hypergeometric', N.units = 1000)
+  prior <- auditPrior(method = 'none', likelihood = 'hypergeometric', N.units = 1000)
   jfaRes <- planning(conf.level = 0.95, materiality = 0.02, expected = 0, likelihood = "hypergeometric", prior = prior, N.units = 1000)
   expect_equal(jfaRes[["posterior"]][["hypotheses"]]$bf.hmin, 1247.05, tolerance = 0.001)
 })
 
 test_that(desc = "(id: f5-v0.4.0-t2) Expected Bayes factors for expected errors > 0", {
-  jfaRes <- planning(conf.level = 0.95, materiality = 0.02, expected = 0.01, likelihood = "poisson", prior = TRUE)
+  prior <- auditPrior(method = 'strict', likelihood = 'poisson')
+  jfaRes <- planning(conf.level = 0.95, materiality = 0.02, expected = 0.01, likelihood = "poisson", prior = prior)
   expect_equal(jfaRes[["posterior"]][["hypotheses"]]$bf.hmin, 19.01191777, tolerance = 0.001)
   
-  prior <- auditPrior(method = 'uniform', likelihood = 'binomial')
+  prior <- auditPrior(method = 'none', likelihood = 'binomial')
   jfaRes <- planning(conf.level = 0.95, materiality = 0.02, expected = 0.01, likelihood = "binomial", prior = prior)
   expect_equal(jfaRes[["posterior"]][["hypotheses"]]$bf.hmin, 933.3458, tolerance = 0.001)
   
-  prior <- auditPrior(method = 'uniform', likelihood = 'hypergeometric', N.units = 1000)
+  prior <- auditPrior(method = 'none', likelihood = 'hypergeometric', N.units = 1000)
   jfaRes <- planning(conf.level = 0.95, materiality = 0.02, expected = 0.01, likelihood = "hypergeometric", prior = prior, N.units = 1000)
   expect_equal(jfaRes[["posterior"]][["hypotheses"]]$bf.hmin, 1619.812, tolerance = 0.001)
 })
@@ -206,7 +208,7 @@ test_that(desc = "(id: f5-v0.5.0-t2) Test for Bayesian summary and print functio
   invisible(capture.output(summary(jfaRes)))
   invisible(capture.output(print(jfaRes[["posterior"]])))
   invisible(capture.output(summary(jfaRes[["posterior"]])))
-  expect_equal(jfaRes[["n"]], 300)
+  expect_equal(jfaRes[["n"]], 299)
   expect_equal(jfaRes[["x"]], 0)
 })
 
@@ -254,7 +256,7 @@ test_that(desc = "(id: f5-v0.5.2-t1) Test for change in Hypergeometric mode calc
 })
 
 test_that(desc = "(id: f5-v0.5.2-t2) Test for change in beta-binomial mode calculation", {
-  prior <- auditPrior(method = 'uniform', likelihood = 'hypergeometric', N.units = 10000)
+  prior <- auditPrior(method = 'none', likelihood = 'hypergeometric', N.units = 10000)
   jfaRes <- planning(materiality = 0.05, conf.level = 0.95, expected = 0, likelihood = "hypergeometric", N = 10000, prior = prior)
   modeDist <- ceiling((jfaRes[["ub"]] - jfaRes[["precision"]]) * 10000)
   expect_equal(jfaRes[["n"]], 58)